ci: sync every week
The upstream sync PR normally includes conflicts so we have to resolve them.
And it requires review before merge so it will take a few days to be merged.
So daily force-sync workflow overwrites our fixes every day and it requires us
to resolve and review and merge in one day.
This is not practical and annoying surely.
As long as we have dispatch workflow enabled, weekly sync is enough.